Create substr_example
This demonstrates how to use substring to select a new column using pre-existing data. In this instance, I have selected 'country'. I wanted the first three letters so the 1 denominates where in the word my substring I am starting, while the 3 denominates the number of characters from the start point. 

AS creates this as a new column in the select, so if I wanted the full name in there for some reason I could do so.